+++ front/CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+## 2.28.0
 ### Added
 - Add container swarm state and a progress bar for container startup
 - Add a button to delete a container if it is KO